Move all other non-blog related assets to index. Update scripts to handle arguments for running server in test and debug mode.
