+++ index/fixtures/admin_interface_theme_prokope.json
