+++ index/templates/search_results.html
